dec/requirements.txt
black
flake8
GitPython
ipython
jupyter
matplotlib
numpy
pandas
pdbpp
pygame
pylint
pytest
seaborn
